Revert "Use impulse strategy in VelocityTracker"
This reverts commit 4f3702edabc0d495db2580916785757365a761da.

Reason for revert: An important application is seeing revenue decrease and this CL was identified as a possible cause. Reverting this will be an experiment to see if the numbers return. If so, we will have to dig into how this CL can cause the problem. If not, we will bring it back again.

Test: ran modified tests

Change-Id: I6c6fdad342688a8ca162d90614a19cb0b12e69b6