Remove --offline from the sandbox build
With Kotlin 1.7.0, Kotlin Gradle Plugin inherits the --offline
flag for its `airplaneMode`.

Unfortunately, with that flag, it doesn't consider any of its
repository parameters for pulling dependencies, hence it fails.

This CL works around that issue by removing the `--offline` flag
from the sandbox build. Note that the sandbox build runs in nsjail
hence it cannot be making any network requests so this is a rather
safe change.

Ideally, we need to work with a proper long term solution but
this is a quick band-aid to get the build working again.

Bug: https://youtrack.jetbrains.com/issue/KT-49247
Bug: 236115214
Test: CI (abdt run L41300000955135022)

Change-Id: Ie7bd8461c0eb4df6a74e5badae4b8f8d75c8bca2