Add experimental ColorRamps to RangedValue complications
Note the full range of Android ColorGradients are not supported,
which is why we've added a new type.

Bug: 236577375
Test: New unit tests
Relnote: We've added experimental support for optional ColorRanges to RangedValueComplicationData. Normally complications would be rendered in colors of the watch face's choosing, but sometimes the ComplicationDataSource is best placed to set the colors e.g. when they have a particular semantic meaning.  E.g. red to blue for temprature.
Change-Id: I5153a3f4839ed70fb3ad6c60c6121810c3c26748