Implement ImageCaptureScreen
Add ImageCaptureScreenStateHolder to handle the states, use cases and
CameraX implementation details. Update ImageCaptureScreen Composable to
display PreviewView and CameraControl buttons.

Add Flash for ImageCapture

The State Holder will check if the Camera supports flash. If it does
not, a placeholder will be rendered. If flash is supported, a toggle is provided to switch the flash modes.
The State Holder will also decide which icon to render based on the
flash mode.

Bug: 234080273
Test: n/a
Change-Id: I23c6e861a97f39947e86adb005a1b5497e6bb241