Fix clipped AM/PM CompactChip on 12-Hour Picker Demo
Replaced the CompactChip with an extra AM/PM Picker. Tested for large and small screen to ensure that the minutes Picker is centered.

Test: androidx.wear.compose.integration.tests
Bug: 236948604
Relnote: "We have fixed some clipping on the Picker 12-hour time picker moving the AM/PM selection from a CompactChip at the top of the screen to a Picker."
Change-Id: I84de84ed18003e0cd766431b4d2bfc24818fc75e