Add new tests for generic and inherited types.
This CL adds tests for coverage of type names of generic and inherited
types in different scenarios. The main purpose of the test is to ensure
that XProcessing is reporting the same type name for both KSP and KAPT.
This can be a bit tricky because it means we have to match types in
kotlin source files with the type names that KAPT generates for those
source files.

Note: there are currently cases where the type name differs between KSP
and KAPT. This CL isn't meant to fix those cases but simply add coverage
for them. They will be fixed in a separate CL.

Bug: 237280547
Test: TypeInheritanceTest
Change-Id: I316954907e549296944321e684055486856082a0