Fix use of MainThread annotations in MediaRouter
Also add documentation to API methods.

Before this change, not all methods that check for the
calling thread have the MainThread annotation, or javadoc
for that, but some do. After this change, all methods that
check for the calling thread (being the main thread) have
the MainThread annotation, and those that belong in the
public API also document it in the javadoc.

Relnote: Fix MainThread annotations in MediaRouter.
Bug: 205124386
Test: Non functional change.
Change-Id: I3ef6ea1113fcd88b33900f790e0f1979901d4cd7