Change a11y announcement for third state of tri-state checkboxes
The third state of tri-state checkboxes is called "Indeterminate" in
Material and "Mixed" in Aria. It's typically displayed like "-", and
used in "parent" checkboxes when only a portion of its child checkboxes
are checked.

For checkboxes with aria-checked="mixed" on web, the default
announcement is usually "Mixed", but certain screen readers also
synthesize it to "Partially checked" which is also correct.
Compose should adopt one of these for consistency.  Personally, I feel
that "Partially checked" is a more intuitive description.

Note that this Compose-controlled string is a fallback because the
AccessibilityNodeInfo API is two-state.  If and when Android introduces
a tri-state checkbox API (b/230363898), TalkBack will control the
announcement, and this string will only remain for legacy purposes.

Bug: 232966113
Test: TriStateCheckboxSample
Change-Id: I5610c3672eb09a807b0c3ad908f329e393e05101