Fix AndroidViewModelFactory back compat
Currently, if you create a custom AndroidViewModelFactory while using
Lifecycle 2.4+, and use the stateful constructor, the create() function
without extras will never get called. This means that you are forced to
override the create function that includes extras, which will break
people that are just converting. The create function should continue to
be called just like it was previously.

This change makes sure that when the create function function with
extras is called, if the the stateful constructor was used we properly
fall back to the create function without extras.

RelNote: "Custom implementations of `AndroidViewModelFactory` now
correctly call the `create(modelClass)` function when using the stateful
constructor with `Lifecycle` 2.4+"
Test: added test
Bug: 238011621

Change-Id: I5b315ae80d1498b92e412d9316a90e7fa0e9fb3a