Make checkArtifactTask configuration cache compatible
checkArtifactTask relied on calling project.tasks during the execution
of this task as it was part of the from {} lambda. This change moves
to using task provider and a map to fix this. This also allows us to
remove explcit dependency on publishMavenPublicationToMavenRepository
as we get that through the new map.

Test: ./gradlew room:room-compiler:checkArtifactTask -> no warnings
Change-Id: I80bc0096549977c56c028b8756e362fd700c47fd