Don't return result when throwing with no lifecycle
We fixed an issue with the ActivityResultRegistry still returning a
result after an error was thrown, but it was only fixed when you
register with a LifecycleOwner. We should fix it for the case when you
don't register with the LifecycleOwner as well.

RelNote: "The `ActivityResultRegistry` will no longer return a result to
the `ActivityResultCallback` when the `launch` call throws any
`Exception` and the callback was registered without a `LifecycleOwner`."
Test: added test
Bug: 238350794

Change-Id: Ia7ff773da287b2d97490bcc553f90020ea9d6067