Add project dependency constraint between paging-common and paging runtime
Added bi-directional project version constraint between paging-common and paging-runtime. If both artifacts are in the dependency tree,
they would have to be on the same version. This is enforced by gradle automatically bumping up the version to meet constraint.

Before constraint added, with direct dependency on paging-runtime:3.1.1 and paging-common:3.2.0-alpha02
+--- androidx.paging:paging-runtime:3.1.1
|    +--- androidx.paging:paging-common:3.1.1 -> 3.2.0-alpha02

After constraint with direct dependency on paging-runtime:3.1.1 and paging-common:3.2.0-alpha02
+--- androidx.paging:paging-runtime:3.1.1 -> 3.2.0-alpha02
|    +--- androidx.paging:paging-common:3.2.0-alpha02

Test: n/a
Bug: 235256201
Change-Id: I0eb1f099edc626519fa3468e428313b185d105c9