Correct the size of the toggle control area in the ToggleChip
Correct the size of the toggle control area of the ToggleChip to match the UX spec. UX Spec calls for a 4.dp spacer and a 24x24.dp toggle control icon area giving 28.dp width. However the implementation is incorrectly giving 36x24.dp toggle control area. This results in taking away 8.dp of usable text label area.

Bug: 240548670
Test: ./gradlew :wear:compose:compose-material:connectedCheck --info --daemon
RelNote: "We have corrected the size of the toggle control area of the ToggleChip as it was not matching its UX spec. The UX Spec calls for a 4.dp spacer between the label and a 24x24.dp toggle control icon area giving a total 28.dp width. However the implementation is incorrectly giving 36x24.dp toggle control area. This results in taking away 8.dp of usable text label area. NOTE: This bug fix gives additional space for the text label and as result can (positively) impact the text layout for overflowing text. It is possible that if you have screenshot tests they may need to be regenerated."
Change-Id: I514c858bd97dfedf860ac818dc27530a4a9e1a81