Split RangedValueComplicationData
Bug: 240572135
Test: Added new tests
Relnote: We have removed the experimental ValueType from RangedValueComplicationData and instead introduced experimental DiscreteRangedValueComplicationData which is like RangedValueComplicationData except for integer range & value. We've also introduced experimental GoalProgressComplicationData which is similar to RangedValueComplicationData except it's for progress towards a goal where min implicitly is zero, and the value is allowed to be larger than targetValue. Note for all RangedValue variants at least one of monochromeImage, text or title must be speciifed.
Change-Id: I9590c240808a1c6f6235e904af403565e1c2efdb