Slider using Layout
Using Layout for our Slider that places the thumb and track relatively. Making our default thumb and track public. Adding sample for slider with custom thumb and slider with custom track and thumb. Also adding some comments and temporary track and thumb methods for RangeSlider to indicate a need to redo RangeSliderImpl to also use a Layout solution.

Bug: 239692491
Test: N/A
Relnote: Adding two overloaded methods for the current Slider API to allow users to pass in a thumb or track to populate the slider.
Change-Id: I21c0022c3882f896f1d4ca855a24cec5d317f641