Handle unknown files when reporting test failures.
As best as I can tell this only happens when the cause of the failure is
a test throwing an exception that is not caught. Previously in this case
junit-gtest would crash while trying to print the test location. This
will cause `unknown file:-1` to print in that case instead.

Bug: None
Change-Id: I7561f1e418298d5d7e1726d0e89ad1cd976af32c
Relnote: junit-gtest no longer crashes when tests don't catch exceptions
Test: added test