Create script to generate offline version of Dackka docs
Partially based on existing development/offlinifyDocs/offlinify_docs.py

Works either for all libraries, or for just one library. Converts links to relative paths, adds CSS, and removes book.yaml metadata. If docs are just for one library, links to pages outside the library point to the published pages, and the main class and package index pages are filtered to just classes and packages in the one library.

Test: `python3 test_offlinify_dackka_docs.py`. Docs generated with `python3 offlinify_dackka_docs.py --library appcompat` are at go/sampleofflinedackkadocs
Bug: 237085937
Change-Id: I701ae34a1f68d262f3c424f13b4b2daca662eb17