Introduce DrawStyle in TextStyle and SpanStyle to enable outlined text drawing.
Bug: b/155421273

This change adds Experimental DrawStyle attribute to both TextStyle and SpanStyle. DrawStyle is a sealed class that effectively corresponds to `android.graphics.Paint.Style` enum which indicates how to draw a path or shape. Until now the default value have been `FILL` which colors in the text as expected. `Stroke` DrawStyle enables developers to color only the outline of a path or text.

Test: :compose:ui:ui-text:testDebug
Test: :compose:ui:ui-text:cAT
Relnote: "Added DrawStyle as an Experimental attribute to TextStyle and SpanStyle to enable drawing outlined text."
Change-Id: If24b8510daa597283c9f4d39c590747ebd822d69