Add way to specify global config for WebView init.
This adds ProcessGlobalConfig class that allows clients to configure
the webview startup parameters prior to loading webview.
The config is stored in ProcessGlobalConfig#sProcessGlobalConfig and
WebView implementation in chromium reflects into this class
to retrieve the config during init.

We are adding setDataDirectorySuffix(String) to be the first config
feature, more will come in later CLs.

The framework dependant subclass of ApiFeature was final to enforce the
mapping between the subclass and the isSupportedByFramework() method
which checks for the corresponding framework version.
This is no longer final,
because it is necessary to extended them to override
isSupportedByWebview() in some cases. To enforce the mapping, the
isSupportedByFramework() method is marked final.

Test: Will be added later once APIs are unhidden
Change-Id: I32ad605d4656812300fa82a495a995da2a91ef1d