Run GlanceAppWidget compositions in SessionWorker
This change updates the glance-appwidget library to use Glance
SessionManager/SessionWorker to run compositions in WorkManager. This
behavior is off by default, but can be turned on by a flag in the source
or in tests.

AppWidgetSessionTest and SizeBoxTest unit tests are added.
GlanceAppWidgetReceiverTest is modified to run all its tests both with
the new SessionWorker and with the current method.

AppWidgetHostRule now waits for Activity.onConfigurationChange when
testing orientation changes.

Bug: 239747024
Test: gradlew :glance:glance-appwidget:{connectedAndroidTest,test}
Relnote: N/A
Change-Id: I6e176d2cd909f0ff5ced48d8f6427e110733d042