Deprecate CameraPipe.External
This change begins the deprecation process for CameraPipe.External
by creating a method that allows a developer to create a CameraGraph
instance by providing a fully specified CameraBackend as part of
the CameraGraph.Config object.

There are two main benefits from this:

1. Both internal/external CameraGraphs are initialized using the
   same set of objects and classes, allowing caches to be reused.
2. More classes can be reused when creating a CameraGraph, which
   reduces code size, and reduces the risk a developer changes
   the behavior of a CameraGraph.

The two main public interfaces this change introduces are the
CaptureSequence and CaptureSequenceProcessor, which replaces the
former RequestProcessor interface. The primary intent of this change
is to allow more logic to be reused between building and submitting
requests to the underlying camera.

This change also has a large number of secondary improvements that
ended up being required as part of refactoring the core APIs. These
include:

* Replacing runBlocking and runBlockingTest with runTest in tests.
* Adding testOnlyDispatcher and testOnlyScope to the threads module.
* Enforcing that several tests use dispatchers based on runTest
* Replacing several test usages of GlobalScope with runTest and launch

Test: ./gradlew --strict\
 :camera:camera-camera2-pipe:testDebugUnitTest\
 :camera:camera-camera2-pipe-testing:testDebugUnitTest\
 :camera:camera-camera2-pipe-integration:testDebugUnitTest

Change-Id: I2776f045312b04522e45e0faa971944988d5da29