Throw descriptive exception when JS sandbox not supported
Makes JavaScriptSandbox.createConnectedInstanceAsync() throw an
exception when the WebView is too old to support JavaScriptSandbox (or
does not exist), and introduces a public isSupported() method.

JavaScriptSandbox is considered supported when the OS has an API level
of at least 26 (enforced by existing minSdkVersion) and WebView is
(around) release M101 or later.

The new SandboxUnsupportedException is an unchecked exception, and is
thrown synchronously with the createConnectedInstanceAsync() call.

The version check is intentionally not performed and the exception not
thrown for createConnectedInstanceForTestingAsync().

Bug: 1354157
Change-Id: I0dcd6f577ce55fdede9cad8c3dfb48905169f8e5
Chrome-Cl: https://crrev.com/c/3838596
GitOrigin-RevId: e6bcbe24214f477a6cd62260f5d27723cb9fa9c9 PatchSet 11
Test: ./gradlew :javascriptengine:javascriptengine:connectedAndroid
Test: Confirmed all skipped when using WebView versions 4951.23, 4952.0
Test: Confirmed partial skipping when using WebView versions 4951.24, 4951.74, 4976.0
Test: Confirmed none skipped when using WebView version 5249.8
Relnote: Add JavaScriptSandbox#isSupported for checking whether the system supports JavaScript sandboxes.
Relnote: Throw SandboxUnsupportedException when JavaScriptSandbox cannot be created due to lack of system support.