Fix navigating to Activities from unloaded modules
Currently, if you are using dynamic Navigation and you attempt to
navigate to a Activity in a module that has not yet be loaded, a
ClassNotFoundException is thrown. This is because the
DynamicActivityNavigator currently attempts to install the module (which
is an async operation) and then immediately navigates to the new class.

Instead, we should return after the install attempt and then allow the
systme to retry the navigation at the appropriate time.

RelNote: "Dynamic Navigation now properly attempts to install Activity
destinations from other modules before navigating to them."
Test: tested on a device using bundletool
Bug: 240292838

Change-Id: Ia2c1645426d2f6a5958a10379a99f2aade3dd03a