Fix TextOverflow.Visible issue with the Canvas override
As discussed, we need to "trick" the android.text.Layout#draw method so that the clipBounds that it reads from Canvas were large enough for it to stop cutting the lines.

This version returns the infinite vertical clipBound. Ideally, we want to restrict this to the display size so we don't overrasterize. Follow-up CL is attempting to do so.

Bug: 240931064
Test: new tests and demo
Change-Id: I4f87e560e660d218495eade66b31e41ea0358a02