[Extensions] Fixed surface abandoned crash when pause/resume too quickly on Samsung Z Fold2
On Samsung Z Fold2, when Extensions are enabled, if pause / resume
camera too quickly, it might cause surface abandoned crash.

Samsung will start the initialization at onEnableSession,
and release the resource at onDisableSession. The issue is
caused by onDisableSession being invoked too early (within
50ms) and could destroy the ImageWriter surface while it
is still used in the initialization.

The workaround is to ensure onDisableSession being called in
at least 100ms after onEnableSession. It will sleep the
camera thread for a while to achieve this. Though the CL
could cause some delay when shuting down the camera,
it is not the normal user operation to close the camera
and open the camera quickly within 100ms. This normally
happens in auto testing. So the impacts for the user
is almost none.

The mechanism is enabled only when Extensions are used.
For regular camera session, there will be no delay.

Bug: 245226085
Test: LifecycleStatusChangeStressTest
Relnote: "Fixed a crash that happens when Extensions are enabled
on Samsung Z Fold2 device and user pause and resume the activity
(or fragment) too quickly."

Change-Id: I005ad1070a259aa2b88d560eed15ced12656b230