Add non-null check for current templateWrapper
There are crashes happening that it tries to use last template before calling onGetTemplate the first time, so the last template wrapper being set up. At this time, mTemplateWrapper is null and will cause crash on app.
This fix will help prevent the crash.
Notice that when this situation happens, instead of using previous template ID (which does not exist yet), it will create a new wrapper with new ID. This will cause the add of step count which may not be added before. Becuase of that, it may cause app to crash when it reaches the step count limit without it being correctly reset.

Test: build, add unittest,  manual run with showcase app and test
bug: b/244327903

Change-Id: I667f33da17cdad330d4b280c55fdce23ed35cfbc