Change "FullyLoaded" to "FullyDrawn"
Fixes: 246642390

reportFullyDrawn() is a well-used API in Activity and
changing the usage to be named related to "loaded"
didn't make sense. This changes the name to be closer
to the "fully drawn."

Also changed the notification to happen on the same
frame that the report was made. An OnPreDrawListener
is used on API 16+ that runs for the first 10 seconds
of the application's life. If it doesn't report
by then, it is removed and postAtFrontOfQueue()
is used instead. For API 15 and earlier,
postAtFrontOfQueue() is used always.

Relnote: "Changed the name of ReportFullyLoaded to
ReportFullyDrawn and other similar name changes."

Test: ran existing (renamed) tests
Change-Id: I99743f32f813a186cb6669a8ed65ca3f62bf3405